# docs/swagger.yaml
definitions:
...
  BankAccount:
    properties:
      accountNumber:
        type: string
      address:
        type: string
      bankName:
        type: string
      currency:
        $ref: '#/definitions/Currency'
      currencyID:
        type: string
      id:
        type: string
      intermediaryBankAddress:
        type: string
      intermediaryBankName:
        type: string
      name:
        type: string
      ownerName:
        type: string
      routingNumber:
        type: string
      swiftCode:
        type: string
      ukSortCode:
        type: string
    type: object
...
  ClientContactInfo:
    properties:
      emails:
        items:
          type: string
        type: array
      id:
        type: string
      isMainContact:
        type: boolean
      name:
        type: string
    type: object
  ClientInfo:
    properties:
      clientAddress:
        type: string
      clientCompany:
        type: string
      contacts:
        items:
          $ref: '#/definitions/ClientContactInfo'
        type: array
    type: object
  CompanyContactInfo:
    properties:
      address:
        type: string
      phone:
        type: string
    type: object
  CompanyInfo:
    properties:
      description:
        type: string
      id:
        type: string
      info:
        additionalProperties:
          $ref: '#/definitions/CompanyContactInfo'
        type: object
      name:
        type: string
      registrationNumber:
        type: string
    type: object
...
  Currency:
    properties:
      id:
        type: string
      locale:
        type: string
      name:
        type: string
      symbol:
        type: string
      type:
        type: string
    type: object
...
  ErrorResponse:
    properties:
      error:
        type: string
      errors:
        items:
          $ref: '#/definitions/view.ApiError'
        type: array
    type: object
...
  Invoice:
    properties:
      bankID:
        type: string
      cc:
        items:
          type: string
        type: array
      conversionAmount:
        type: number
      conversionRate:
        type: number
      description:
        type: string
      discount:
        type: number
      dueAt:
        type: string
      email:
        type: string
      errorInvoiceID:
        type: string
      failedAt:
        type: string
      invoiceFileURL:
        type: string
      invoicedAt:
        type: string
      lineItems:
        items:
          $ref: '#/definitions/github.com_dwarvesf_fortress-api_pkg_view.InvoiceItem'
        type: array
      month:
        type: integer
      note:
        type: string
      number:
        type: string
      paidAt:
        type: string
      projectID:
        type: string
      scheduledDate:
        type: string
      sentBy:
        type: string
      status:
        type: string
      subTotal:
        type: number
      tax:
        type: number
      threadID:
        type: string
      total:
        type: number
      year:
        type: integer
    type: object
  InvoiceData:
    properties:
      bankAccount:
        $ref: '#/definitions/BankAccount'
      bankID:
        type: string
      cc:
        items:
          type: string
        type: array
      client:
        $ref: '#/definitions/ClientInfo'
      companyInfo:
        $ref: '#/definitions/CompanyInfo'
      conversionAmount:
        type: number
      conversionRate:
        type: number
      description:
        type: string
      discount:
        type: number
      dueAt:
        type: string
      email:
        type: string
      errorInvoiceID:
        type: string
      failedAt:
        type: string
      invoiceFileURL:
        type: string
      invoicedAt:
        type: string
      lineItems:
        items:
          $ref: '#/definitions/github.com_dwarvesf_fortress-api_pkg_view.InvoiceItem'
        type: array
      month:
        type: integer
      note:
        type: string
      number:
        type: string
      paidAt:
        type: string
      projectID:
        type: string
      projectName:
        type: string
      scheduledDate:
        type: string
      sentBy:
        type: string
      status:
        type: string
      subTotal:
        type: number
      tax:
        type: number
      threadID:
        type: string
      total:
        type: number
      year:
        type: integer
    type: object
  InvoiceListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/InvoiceData'
        type: array
      page:
        type: integer
      size:
        type: integer
      sort:
        type: string
      total:
        type: integer
    type: object
  InvoiceTemplateResponse:
    properties:
      data:
        $ref: '#/definitions/ProjectInvoiceTemplate'
    type: object
...
  MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  ProjectInvoiceTemplate:
    properties:
      bankAccount:
        $ref: '#/definitions/BankAccount'
      client:
        $ref: '#/definitions/ClientInfo'
      companyInfo:
        $ref: '#/definitions/CompanyInfo'
      id:
        type: string
      invoiceNumber:
        type: string
      lastInvoice:
        $ref: '#/definitions/Invoice'
      name:
        type: string
    type: object
...
  SendInvoiceRequest:
    properties:
      bankID:
        type: string
      cc:
        items:
          type: string
        type: array
      description:
        type: string
      discount:
        minimum: 0
        type: number
      dueDate:
        type: string
      email:
        type: string
      invoiceDate:
        type: string
      invoiceMonth:
        maximum: 11
        minimum: 0
        type: integer
      invoiceYear:
        minimum: 0
        type: integer
      isDraft:
        type: boolean
      lineItems:
        items:
          $ref: '#/definitions/github.com_dwarvesf_fortress-api_pkg_handler_invoice_request.InvoiceItem'
        type: array
      note:
        type: string
      number:
        type: string
      projectID:
        type: string
      subtotal:
        minimum: 0
        type: number
      tax:
        minimum: 0
        type: number
      total:
        minimum: 0
        type: number
    required:
    - bankID
    - dueDate
    - email
    - invoiceDate
    - projectID
    type: object
...
  github.com_dwarvesf_fortress-api_pkg_handler_invoice_request.InvoiceItem:
    properties:
      cost:
        type: number
      description:
        type: string
      discount:
        type: number
      isExternal:
        type: boolean
      quantity:
        type: number
      unitCost:
        type: number
    type: object
  github.com_dwarvesf_fortress-api_pkg_view.InvoiceItem:
    properties:
      cost:
        type: number
      description:
        type: string
      discount:
        type: number
      isExternal:
        type: boolean
      quantity:
        type: number
      unitCost:
        type: number
    type: object
...
  view.ApiError:
    description: validation error details
    properties:
      enums:
        description: available options incase of field's payload is enums
        items:
          type: string
        type: array
      field:
        description: the field cause the error
        type: string
      msg:
        description: error message
        type: string
    type: object
...
info:
  contact:
    email: benjamin@d.foundation
    name: Nam Nguyen
    url: https://d.foundation
  description: This is api document for fortress project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FORTRESS API DOCUMENT
  version: v0.1.39
paths:
...
  /invoices:
    get:
      consumes:
      - application/json
      description: Get latest invoice by project id
      operationId: getInvoiceList
      parameters:
      - description: projectID
        in: query
        name: projectID
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: sort
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/InvoiceListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get latest invoice by project id
      tags:
      - Invoice
  /invoices/{id}/status:
    put:
      consumes:
      - application/json
      description: Update status for invoice
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update status for invoice
      tags:
      - Invoice
  /invoices/send:
    post:
      consumes:
      - application/json
      description: Create new invoice and send to client
      operationId: sendInvoice
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SendInvoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new invoice and send to client
      tags:
      - Invoice
  /invoices/template:
    get:
      consumes:
      - application/json
      description: Get the latest invoice by project id
      operationId: getInvoiceTemplate
      parameters:
      - description: projectID
        in: query
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/InvoiceTemplateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the latest invoice by project id
      tags:
      - Invoice
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"